refactor: FormDetailPage renamed as FormFilloutPage
