+++ src/features/form/components/index.ts
@@ -1,7 +1,7 @@
 export { AdminFormDetailPage } from "./AdminFormDetailPage";
 export { AdminFormPreviewPage } from "./AdminFormPreviewPage";
 export { AdminFormsPage } from "./AdminFormsPage";
-export { FormDetailPage } from "./FormDetailPage";
 export { FormEntryPage } from "./FormEntryPage";
+export { FormFilloutPage } from "./FormFilloutPage";
 export { FormsListPage } from "./FormsListPage";
 export { OAuthConnectCallbackPage } from "./OAuthConnectCallbackPage";
